create Location struct and change how location update notification works
