+++ Cloudy Nine/LocationManager.swift
@@ -9,6 +9,12 @@
 import Foundation
 import CoreLocation
 
+struct Location {
+    let lat: Double
+    let lon: Double
+    let city: String?
+}
+
 protocol LocationManagerPermissionDelegate {
     func userDidGrantLocationPermission()
 }
@@ -18,6 +24,7 @@ class LocationManager: NSObject {
     // MARK: Private properties
 
     private let manager = CLLocationManager()
+    private let geocoder = CLGeocoder()
 
     // MARK: Public properties
 
@@ -26,9 +33,9 @@ class LocationManager: NSObject {
     var permissionGranted: Bool {
         return CLLocationManager.authorizationStatus() == .AuthorizedWhenInUse
     }
+    var location: Location?
 
     static let LocationUpdateNotificationName = "LocationManagerLocationUpdateNotification"
-    static let LocationUpdateNotificationLocationKey = "LocationUpdateNotificationLocation"
 
     // MARK: Initialization
 
@@ -57,7 +64,21 @@ class LocationManager: NSObject {
     func requestAuthorization() {
         manager.requestWhenInUseAuthorization()
     }
+}
 
+// MARK: Private implementation
+
+private extension LocationManager {
+    func cityFromLocation(location: CLLocation, completion: (String?) -> Void) {
+        geocoder.reverseGeocodeLocation(location) { placemarks, error in
+            guard let placemark = placemarks?.first else {
+                completion(nil)
+                return
+            }
+
+            completion(placemark.locality)
+        }
+    }
 }
 
 // MARK: CLLocationManagerDelegate
@@ -77,12 +98,16 @@ extension LocationManager: CLLocationManagerDelegate {
             return
         }
 
-        let notification = NSNotification(
-            name: LocationManager.LocationUpdateNotificationName,
-            object: nil,
-            userInfo: [LocationManager.LocationUpdateNotificationLocationKey: newestLocation]
-        )
+        cityFromLocation(newestLocation) { city in
+            self.location = Location(lat: newestLocation.coordinate.latitude, lon: newestLocation.coordinate.longitude, city: city)
 
-        NSNotificationCenter.defaultCenter().postNotification(notification)
+            let notification = NSNotification(
+                name: LocationManager.LocationUpdateNotificationName,
+                object: nil,
+                userInfo: nil
+            )
+
+            NSNotificationCenter.defaultCenter().postNotification(notification)
+        }
     }
 }
